requirements.txt
open ai version==0.28.0
python-dotenv
flask